refactor: move orders stepper into orders domain
