+++ src/components/orders/stepper/StepPickupConfirmation.tsx
@@ -2,9 +2,9 @@ import React, { useContext } from "react";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import { Typography } from "@mui/material";
-import { stepperContext } from "../../providers/StepperProvider";
+import { stepperContext } from "../../../providers/StepperProvider";
 
-const StepPickup = () => {
+const StepPickupConirmation = () => {
     const { activeStep, setActiveStep } = useContext(stepperContext);
     const handleSetActiveStep = (prevActiveStep: number) => {
         return prevActiveStep + 1;
@@ -16,13 +16,13 @@ const StepPickup = () => {
             <Box>
                 <Box sx={{ display: "flex", flexDirection: "column", justifyContent: "center", alignItems: "center", pt: 10 }} >
                     <Typography sx={{ pb: 4 }} >
-                        New Order Added for:
+                        Hand off:
                     </Typography>
                     <Typography sx={{ pb: 4 }}>
-                        John Doe
+                        John Doe : 3 boxes
                     </Typography>
                     <Typography sx={{ pb: 4 }}>
-                        Order No:
+                        AFAC: 3 boxes
                     </Typography>
                 </Box>
                 <Box display="flex"
@@ -36,7 +36,8 @@ const StepPickup = () => {
                         color='secondary'
                         onClick={handlePickup}
                     >
-                        <Typography variant='button'>Proceed To Pick Up</Typography>
+                        <Typography variant='button'>Mark as picked up
+                        </Typography>
                     </Button>
                 </Box>
             </Box>
@@ -44,4 +45,4 @@ const StepPickup = () => {
     );
 };
 
-export default StepPickup;
+export default StepPickupConirmation;
